requirements.txt.txt
asgiref==3.8.1
attrs==25.3.0
branca==0.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
geographiclib==2.0
geopy==2.4.1
h11==0.16.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.6
outcome==1.3.0.post0
packaging==25.0
pillow==11.2.1
pycparser==2.22
PyJWT==2.9.0
PySocks==1.7.1
python-dotenv==1.1.0
requests==2.32.3
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
websocket-client==1.8.0
wsproto==1.2.0
xyzservices==2025.4.0
